ckanext-gobcangroups/requirements.txt
asn1crypto==0.24.0
Babel==2.3.4
Beaker==1.8.1
bleach==1.5.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
-e git+https://github.com/ckan/ckan@20a506ddbce33c92e3dfc510e1f1d7097caa45f9#egg=ckan
ckanext-caopendata==0.0.1
-e git+https://github.com/ckan/ckanext-dcat.git@0d445d31fcb1cec1ccb669ba8350c8fceac20a0e#egg=ckanext_dcat
-e git+https://github.com/damalaga/ckanext-federador@5ccf5e537d33d47fe1a908d1c91ab40faf4fb52e#egg=ckanext_federador
ckanext-gobcangroups==0.0.1
ckanext-gobcanopendata==0.0.1
-e git+https://github.com/ckan/ckanext-harvest.git@b8cf47bb91802540c1ce2cdcc6d313849228fdf3#egg=ckanext_harvest
ckantoolkit==0.0.3
click==6.7
cryptography==2.7
decorator==4.0.6
enum34==1.1.6
fanstatic==0.12
Flask==0.11.1
FormEncode==1.3.0
funcsigs==1.0.2
geomet==0.2.0.post2
html5lib==0.9999999
idna==2.7
ipaddress==1.0.22
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
Markdown==2.6.7
MarkupSafe==0.23
nose==1.3.7
ofs==0.4.2
ordereddict==1.1
Pairtree===0.7.1-T
passlib==1.6.5
Paste==1.7.5.1
PasteDeploy==1.5.2
PasteScript==2.0.2
pbr==1.10.0
pika==0.9.8
polib==1.0.7
psycopg2==2.4.5
pycparser==2.19
Pygments==2.1.3
Pylons==0.9.7
pyOpenSSL==18.0.0
pyparsing==2.2.2
pysolr==3.6.0
python-dateutil==1.5
python-magic==0.4.12
pytz==2016.7
pyutilib.component.core==4.6.4
rdflib==4.2.1
rdflib-jsonld==0.4.0
redis==2.10.1
repoze.lru==0.6
repoze.who==2.3
repoze.who-friendlyform==1.0.8
requests==2.20.0
Routes==1.13
rq==0.6.0
simplejson==3.10.0
six==1.10.0
SPARQLWrapper==1.8.2
SQLAlchemy==0.9.6
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
tzlocal==1.3
unicodecsv==0.14.1
urllib3==1.24.3
vdm==0.13
WebError==0.13.1
WebHelpers==1.3
WebOb==1.0.8
WebTest==1.4.3
Werkzeug==0.11.10
zope.interface==4.3.2
